Basic engine customizations
Nothing lives forever. We don't want pastes to be around till the end of
the world. So from now on, only pastes with limited time to live are
acceptable.

Also disabling features that just doesn't work (or we don't like them
(as flash)).

Everything is by default text, if user doesn't say something else.