Fixed URI checking
In new php bug #47229 “preg_quote should escape “-” (minus) as well” was
fixed. (technically in 5.2.8) sticked checks uri for allowed characters
to prevent some bad things.  But it use preg_quote to convert the
allowed list of character to something usable in a regular expression.
Now the minus “-”, gets escaped in preg_quote with a backslash “\”.
That cause the expression “a-z 0-9″ to be converted to “a\-z 0\-9″ which
will not work in a regex. So removing preg_quote, will fix the problem
and we can use this pastebin even with new php ;-)