Added support for presets through GET
For some use-cases, it might be useful to provide link with some of the
item on the create page prefilled. If these default values can be
provided using GET, it could make it easier to integrate openSUSE Paste
into various application/web pages. Example would be providing url with
shorter ttl for IRC, longer for mailing lists and forums. Or setting
specific title for bug report posts. Or creating a web page that will
pregenerate funny nick names.